Make size run in constant time
This PR modifies the HashMap datatype and its associated functions
so that `size :: (Eq k, Hashable k) => HashMap k v -> Int` runs in
constant time, and not linear.

To summarize what was done (not in this order):

* the present (as of now, before this PR is merged) `HashMap` type was
  renamed Tree and is no longer visible to the end-user;
* a wrapper `HashMap = HashMap !Int !(Tree k v)` was created, is now
  the user-facing `HashMap` datatype. Here, the `!Int` represents the
  hashmap's size;
* every function that potentially changes a hashmap's size calls an
  internal "wrapper" version of itself that returns a tuple with the
  change in the hashmap's size;
* the wrapper function is responsible for updating the `Int` field in
  HashMap after calling its internal variant because of all this, size
  is now defined as `size (HashMap sz _) = sz`, which runs in `O(1)`;

This PR also fixes some typos in comments, and adds some to helper
functions for clarity.

As an added positive change unrelated to the `size` function,
`(==)/(/=)` will now shortcircuit whenever the two hashmaps being
compared don't have the same size, which in practically every case
means equality comparisons run in constant time.